+++ core/src/main/java/io/snabble/sdk/events/Events.kt
@@ -5,9 +5,7 @@ import android.os.Looper
 import android.os.SystemClock
 import io.snabble.sdk.Project
 import io.snabble.sdk.Shop
-import io.snabble.sdk.ShoppingCart
-import io.snabble.sdk.ShoppingCart.SimpleShoppingCartListener
-import io.snabble.sdk.ShoppingCart.Taxation
+import io.snabble.sdk.shoppingcart.ShoppingCart
 import io.snabble.sdk.Snabble
 import io.snabble.sdk.codes.ScannedCode
 import io.snabble.sdk.events.data.Event
@@ -20,6 +18,8 @@ import io.snabble.sdk.events.data.payload.PayloadLog
 import io.snabble.sdk.events.data.payload.PayloadProductNotFound
 import io.snabble.sdk.events.data.payload.PayloadSessionEnd
 import io.snabble.sdk.events.data.payload.PayloadSessionStart
+import io.snabble.sdk.shoppingcart.data.Taxation
+import io.snabble.sdk.shoppingcart.data.listener.SimpleShoppingCartListener
 import io.snabble.sdk.utils.DateUtils
 import io.snabble.sdk.utils.GsonHolder
 import io.snabble.sdk.utils.Logger
@@ -51,7 +51,7 @@ class Events internal constructor(
     init {
         project.shoppingCart.addListener(object : SimpleShoppingCartListener() {
 
-            override fun onChanged(cart: ShoppingCart) {
+            override fun onChanged(cart: ShoppingCart?) {
                 updateShop(Snabble.checkedInShop)
                 shop?.let {
                     if (!hasSentSessionStart) {
@@ -62,29 +62,29 @@ class Events internal constructor(
                 }
             }
 
-            override fun onCleared(cart: ShoppingCart) {
+            override fun onCleared(cart: ShoppingCart?) {
                 val isSameCartWithNewId = shoppingCart === cart && cart.id != cartId
                 if (isSameCartWithNewId) {
                     val payloadSessionEnd = PayloadSessionEnd(session = cartId)
                     post(payloadSessionEnd, false)
-                    cartId = cart.id
+                    cartId = cart?.id
                     hasSentSessionStart = false
                 }
             }
 
-            override fun onProductsUpdated(list: ShoppingCart) {
+            override fun onProductsUpdated(list: ShoppingCart?) {
                 // Override because it shouldn't trigger onChanged(Cart)
             }
 
-            override fun onPricesUpdated(list: ShoppingCart) {
+            override fun onPricesUpdated(list: ShoppingCart?) {
                 // Override because it shouldn't trigger onChanged(Cart)
             }
 
-            override fun onTaxationChanged(list: ShoppingCart, taxation: Taxation) {
+            override fun onTaxationChanged(list: ShoppingCart?, taxation: Taxation?) {
                 // Override because it shouldn't trigger onChanged(Cart)
             }
 
-            override fun onCartDataChanged(list: ShoppingCart) {
+            override fun onCartDataChanged(list: ShoppingCart?) {
                 // Override because it shouldn't trigger onChanged(Cart)
             }
         })