+++ core/src/test/java/io/snabble/sdk/checkout/MockCheckoutApi.kt
@@ -3,9 +3,10 @@ package io.snabble.sdk.checkout
 import com.google.gson.JsonObject
 import io.snabble.sdk.PaymentMethod
 import io.snabble.sdk.Project
-import io.snabble.sdk.ShoppingCart
+import io.snabble.sdk.shoppingcart.ShoppingCart
 import io.snabble.sdk.merge
 import io.snabble.sdk.payment.PaymentCredentials
+import io.snabble.sdk.shoppingcart.data.cart.BackendCart
 import io.snabble.sdk.utils.GsonHolder
 import java.util.*
 
@@ -30,7 +31,7 @@ class MockCheckoutApi(
     }
 
     override fun createCheckoutInfo(
-        backendCart: ShoppingCart.BackendCart,
+        backendCart: BackendCart,
         checkoutInfoResult: CheckoutInfoResult?,
         timeout: Long
     ) {