+++ kotlin-sample/src/main/java/io/snabble/sdk/sample/coupons/ui/widget/CouponHeaderWidget.kt
@@ -1,0 +1,42 @@
+package io.snabble.sdk.sample.coupons.ui.widget
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun CouponHeaderWidget(modifier: Modifier = Modifier, name: String, description: String?) {
+    Column(
+        modifier = modifier,
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        Text(
+            text = name,
+            style = MaterialTheme.typography.titleMedium,
+            fontWeight = FontWeight.Bold
+        )
+        description?.let {
+            Text(
+                text = it,
+                style = MaterialTheme.typography.bodySmall
+            )
+        }
+    }
+}
+
+@Composable
+@Preview(showBackground = true)
+private fun Preview() {
+    CouponHeaderWidget(
+        modifier = Modifier.padding(16.dp),
+        name = "Demo Coupon",
+        description = "Everything is free"
+    )
+}