+++ ui/src/main/java/io/snabble/sdk/ui/cart/deprecated/ShoppingCartFragment.kt
@@ -1,4 +1,4 @@
-package io.snabble.sdk.ui.cart
+package io.snabble.sdk.ui.cart.deprecated
 
 import android.os.Bundle
 import android.view.Menu