+++ ui/src/main/java/io/snabble/sdk/ui/cart/shoppingcart/CartItem.kt
@@ -1,0 +1,8 @@
+package io.snabble.sdk.ui.cart.shoppingcart
+
+import io.snabble.sdk.shoppingcart.ShoppingCart
+
+interface CartItem {
+
+    val item: ShoppingCart.Item
+}