+++ ui/src/main/java/io/snabble/sdk/ui/cart/shoppingcart/product/model/DiscountItem.kt
@@ -1,0 +1,7 @@
+package io.snabble.sdk.ui.cart.shoppingcart.product.model
+
+internal data class DiscountItem(
+    val name: String,
+    val discount: String,
+    val discountValue: Int
+)